Integrate content build script with the app
