+++ app/package.json
@@ -6,7 +6,7 @@
     "scripts": {
         "dev": "svelte-kit dev",
         "dev:host": "svelte-kit dev --host",
-        "build:content": "node ../content/scripts/esbuild.js && node ../content/compiled/build-content.js",
+        "build:content": "node scripts/esbuild.js && node scripts/compiled/build-content.js",
         "build": "svelte-kit build",
         "build:all": "npm run build:content && svelte-kit build",
         "preview": "svelte-kit preview",
@@ -26,20 +26,21 @@
         "@types/sanitize-html": "^2.6.2",
         "autoprefixer": "^10.4.7",
         "cssnano": "^5.1.11",
+        "dotenv": "^16.0.1",
+        "native-fetch": "^4.0.2",
         "postcss": "^8.4.14",
         "postcss-load-config": "^3.1.4",
         "prettier": "^2.7.1",
         "prettier-plugin-svelte": "^2.7.0",
         "prettier-plugin-tailwindcss": "^0.1.11",
         "remark-slate": "^1.8.6",
+        "slugify": "^1.6.5",
         "svelte": "^3.48.0",
         "svelte-check": "^2.7.2",
         "svelte-preprocess": "^4.10.7",
         "tailwindcss": "^3.1.3",
         "tslib": "^2.4.0",
         "typescript": "^4.7.3",
-        "dotenv": "^16.0.1",
-        "native-fetch": "^4.0.2",
         "undici": "^5.5.1"
     },
     "type": "module",
