+++ app/tsconfig.json
@@ -2,7 +2,7 @@
     "extends": "./.svelte-kit/tsconfig.json",
     "compilerOptions": {
         "moduleResolution": "node",
-        "module": "es2020",
+        "module": "esnext",
         "lib": ["es2020", "DOM"],
         "target": "es2020",
         /**
@@ -37,6 +37,7 @@
         "src/**/*.d.ts",
         "src/**/*.js",
         "src/**/*.ts",
-        "src/**/*.svelte"
+        "src/**/*.svelte",
+        "scripts/**/*.ts"
     ]
 }
