--- content/.gitignore
@@ -1,5 +1,0 @@
-.DS_Store
-node_modules
-.env
-
-compiled
