--- content/tsconfig.json
@@ -1,25 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "esnext",
-        "lib": ["esnext"],
-        "allowJs": true,
-        "checkJs": true,
-        "skipLibCheck": true,
-        "esModuleInterop": true,
-        "allowSyntheticDefaultImports": true,
-        "strict": true,
-        "forceConsistentCasingInFileNames": true,
-        "noImplicitAny": true,
-        "module": "esnext",
-        "moduleResolution": "node",
-        "resolveJsonModule": true,
-        "isolatedModules": true,
-        "baseUrl": ".",
-        "paths": {
-            "$shared": ["../shared"],
-            "$shared/*": ["../shared/*"]
-        }
-    },
-    "include": ["scripts/**/*.ts"],
-    "exclude": ["node_modules", "src"]
-}
